MyProject.Test/Application.Test/Services.Test/UserGroupServiceTests.cs
using System.Collections.Generic;
using System.Threading.Tasks;
using Moq;
using MyProject.Application.Interfaces;
using MyProject.Domain.Entities;
using MyProject.Application.Services;
using Xunit;
using MyProject.Application.IRepositores;

namespace MyProject.Tests.Services
{
    public class UserGroupServiceTests
    {
        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetByIdAsync_ValidId_ReturnsUserGroup(int userGroupId)
        {
            // Arrange
            var mockRepository = new Mock<IUserGroupRepository>();
            var userGroupService = new UserGroupService(mockRepository.Object);

            var expectedUserGroup = new UserGroup { Id = userGroupId, Name = "TestGroup" }; // Replace with test data
            mockRepository.Setup(repo => repo.GetByIdAsync(userGroupId)).ReturnsAsync(expectedUserGroup);

            // Act
            var result = await userGroupService.GetByIdAsync(userGroupId);

            // Assert
            Assert.NotNull(result);
            Assert.Equal(userGroupId, result.Id);
            // Add more specific assertions based on your expected behavior
        }

        [Theory]
        [InlineData("Group 1")]
        [InlineData("Group 2")]
        public async Task CreateAsync_ValidUserGroup_ReturnsUserId(string groupName)
        {
            // Arrange
            var mockRepository = new Mock<IUserGroupRepository>();
            var userGroupService = new UserGroupService(mockRepository.Object);

            var createUserGroupRequest = new UserGroup { Name = groupName }; // Replace with test data
            var createdUserGroupId = 1; // Replace with an expected ID for testing purposes
            mockRepository.Setup(repo => repo.AddAsync(createUserGroupRequest)).ReturnsAsync(new UserGroup { Id = createdUserGroupId });

            // Act
            var result = await userGroupService.CreateAsync(createUserGroupRequest);

            // Assert
            Assert.Equal(createdUserGroupId, result);
            // Add more specific assertions based on your expected behavior
        }

        // Similar theories for other methods (GetAllAsync, UpdateAsync, DeleteAsync) with different scenarios

        // Remember to add specific assertions to validate the responses and behaviors of each method.
    }
}
